[MOVE/#3] User 데이터 클래스 domain 패키지로 이동
