+++ app/src/main/java/org/android/go/sopt/domain/repository/AuthRepository.kt
@@ -1,6 +1,6 @@
 package org.android.go.sopt.domain.repository
 
-import org.android.go.sopt.data.entity.User
+import org.android.go.sopt.domain.model.User
 
 interface AuthRepository {
     fun setAutoLogin(isAutoLogin: Boolean)
