+++ app/src/main/java/org/android/go/sopt/presentation/main/profile/ProfileViewModel.kt
@@ -3,7 +3,7 @@ package org.android.go.sopt.presentation.main.profile
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import org.android.go.sopt.data.entity.User
+import org.android.go.sopt.domain.model.User
 import org.android.go.sopt.domain.repository.AuthRepository
 import javax.inject.Inject
