+++ app/src/main/java/org/android/go/sopt/presentation/signup/SignupViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import org.android.go.sopt.data.entity.User
+import org.android.go.sopt.domain.model.User
 import org.android.go.sopt.domain.repository.AuthRepository
 import org.android.go.sopt.util.UiState
 import org.android.go.sopt.util.UiState.Failure
